+++ doweeklycleanup.sh
@@ -13,9 +13,12 @@ current_time=$(date "+%Y.%m.%d-%H.%M.%S")
 zipname=$filedir$current_time.speedtest.zip
 email_addresses="email@email.com email@email2.com"
 
-# First lets cleanup all the weeks text files into a zip
-zip $zipname *.speedtest.txt
-# Now remove all those files
-rm *.speedtest.txt -f
-# Now lets email that attachment so we havew a copy of it
-mpack -s "Weekly Speedtesting Archive" $zipname $email_addresses
+if [ -e "$filedir$filename" ]
+then
+   # First lets cleanup all the weeks text files into a zip
+   zip $zipname *.speedtest.txt
+   # Now remove all those files
+   rm *.speedtest.txt -f
+   # Now lets email that attachment so we havew a copy of it
+   mpack -s "Weekly Speedtesting Archive" $zipname $email_addresses
+fi
